event-listener: fallback for api-service not reachable
