+++ event-listener/src/main/java/com/cyverse/keycloak/portal/service/NoOpUserPortalServiceImpl.java
@@ -1,0 +1,13 @@
+package com.cyverse.keycloak.portal.service;
+
+import org.jboss.logging.Logger;
+import org.keycloak.models.UserModel;
+
+public class NoOpUserPortalServiceImpl implements UserPortalService {
+    private static final Logger logger = Logger.getLogger(NoOpUserPortalServiceImpl.class);
+
+    @Override
+    public void addUserToPortal(UserModel user) {
+        logger.error("No operation possible. Check other dependencies for failures.");
+    }
+}
