+++ register-test/run-tests.sh
@@ -37,7 +37,11 @@ docker exec -i $db_container_id psql -U postgres -c "create database keycloak wi
 docker exec -i $db_container_id psql -U postgres -c "grant postgres to keycloak;"
 
 echo "Waiting for Keycloak to initialize"
-sleep 20
+
+for i in {1..30}; do
+  echo -n "."
+  sleep 1
+done
 
 mvn clean package
